requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
scikit-learn==1.7.2
numpy==1.26.4
pandas==2.2.1
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
python-multipart==0.0.9
jinja2==3.1.3
aiofiles==23.2.1